+++ src/main/java/seedu/address/model/transportbooking/Location.java
@@ -1,4 +1,4 @@
-package seedu.address.model.transportation;
+package seedu.address.model.transportbooking;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
