+++ src/main/java/seedu/address/model/transportbooking/exceptions/TransportBookingNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.transportation.exceptions;
+package seedu.address.model.transportbooking.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified transport booking.
